build: consistently reference the local kind version
The Makefile downloads the defined kind version to bin/kind and
should use this binary when calling kind as kind could be installed
on the OS as well. The kind version used to create the cluster might
not be compatible with the kind version used to e.g. load images.

Consistently reference the locally defined and downloaded kind version
within the Makefile.

Signed-off-by: Aslak Knutsen <aslak@4fs.no>